+++ muk_fields_lobject/doc/changelog.rst
@@ -1,3 +1,11 @@
+`2.0.0`
+-------
+
+- Migrated to Python 3
+- Can handle Base64 input
+- Can return value as Base64
+- Can return value as Checksum
+
 `1.0.0`
 -------